+++ VCPU32-Simulator/VCPU32-DvrExprEvaluator.cpp
@@ -676,7 +676,7 @@ void DrvExprEvaluator::parseExpr( DrvExpr *rExpr ) {
         glb -> tok -> nextToken( );
         parseTerm( rExpr );
         
-        if ( rExpr -> typ == TYP_NUM ) rExpr -> numVal = - rExpr -> numVal;
+        if ( rExpr -> typ == TYP_NUM ) rExpr -> numVal = - (int32_t) rExpr -> numVal;
         else throw ( ERR_EXPECTED_NUMERIC );
     }
     else parseTerm( rExpr );
